+++ ext/PPFQuasiMonteCarloExt.jl
@@ -8,11 +8,11 @@ using QuasiMonteCarlo: SamplingAlgorithm, sample
 ProbabilisticPowerFlow.QuasiMC(sampler::SamplingAlgorithm; n::Integer = 1024) =
     QuasiMC(sampler, n)
 
-function CommonSolve.solve(prob::PPFProblem, method::QuasiMC)
+function CommonSolve.solve(prob::PPFProblem, method::QuasiMC; kwargs...)
     d = germ_dim(prob.model)
     U = Matrix{Float64}(sample(method.n, d, method.sampler))
     clamp!(U, eps(), 1 - eps())
-    return solve_samples(prob, method, U)
+    return solve_samples(prob, method, U; kwargs...)
 end
 
 end